requirements.txt
asgiref==3.6.0
django-cors-headers==3.13.0
Django==4.1.7
djangorestframework==3.14.0
pip==23.0
pytz==2022.7.1
setuptools==65.5.1
sqlparse==0.4.3
wheel==0.38.4
